TO Do List App(html,css,JavaScript)
As a beginner  How To Create TO DO List App Using HTML CSS And JavaScript | Task App In JavaScript Step By Step
In this To Do list we can add our task list and mark that task as completed or incomplete. We can write the new task in the input box and click on the "Add" button then the new task will be added to the list.
We can click on the task name to mark it as completed or incomplete task. We will add a close icon to delete the particular task from the ToDo list.